requirements.txt
langchain==0.1.9
langchain-community==0.0.27
langchain-core==0.1.32
openai==1.12.0
python-dotenv==1.0.1
pytest==8.0.2
rich==13.7.0
typing-extensions==4.9.0 